requirements.txt
mysql-connector==2.2.9
python-dotenv==0.19.2
async-generator==1.10
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
click==8.0.4
cryptography==36.0.1
cx-Oracle==8.3.0
Flask==2.0.3
h11==0.13.0
idna==3.3
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
mysql-connector==2.2.9
numpy==1.22.2
outcome==1.1.0
packaging==21.3
pandas==1.4.1
Pillow==9.0.1
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.7
PySocks==1.7.1
pytesseract==0.3.9
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
selenium==4.1.2
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
trio==0.20.0
trio-websocket==0.9.2
urllib3==1.26.8
Werkzeug==2.0.3
wsproto==1.1.0
async-generator==1.10
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
click==8.0.4
cryptography==36.0.1
cx-Oracle==8.3.0
Flask==2.0.3
Flask-Cors==3.0.10
greenlet==1.1.2
h11==0.13.0
idna==3.3
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
mysql-connector==2.2.9
numpy==1.22.2
outcome==1.1.0
packaging==21.3
pandas==1.4.1
Pillow==9.0.1
pycparser==2.21
pyodbc==4.0.32
pyOpenSSL==22.0.0
pyparsing==3.0.7
PySocks==1.7.1
pytesseract==0.3.9
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
selenium==4.1.2
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.31
trio==0.20.0
trio-websocket==0.9.2
urllib3==1.26.8
Werkzeug==2.0.3
wsproto==1.1.0
